backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.0.0
django-filter==23.2
drf-yasg==1.21.4